% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{add_messy_matrix}
\alias{add_messy_matrix}
\title{The time series tools, for breaking the datset down into different time slices.}
\usage{
add_messy_matrix(m1, m2)
}
\description{
For adding two matrices that might have missing values. Uses a series of 
ifelse statements to add two matrices together that might have missing values.
}
\details{
m1: a matrix that might have missing  values. 
m2: a metrix that might have missing values. 

returns: the sum of the two matrices, where any missing values are dropped if 
     the other matrix has a value. otherwise it's NA
}
