% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maled_tools.R
\name{append_country_vars}
\alias{append_country_vars}
\title{This is the toolkit for the mal-ed dataset.
USed for tools that are specific to the mal-ed set.}
\usage{
append_country_vars(maled, parent_dir = "../data/")
}
\description{
This is the toolkit for the mal-ed dataset.
USed for tools that are specific to the mal-ed set.
}
