% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{atepec_definition}
\alias{atepec_definition}
\title{Set the rules for atEPEC}
\usage{
atepec_definition(bfpa, eae, stx1, stx2, threshold)
}
\value{
A vector of values in range [0,35] with missing values to represent the
column foe atEPEC
}
\description{
Applying the rules provided in 'Ecoli definitions.docx'
for atEPEC and setting as a new column.
}
\details{
Rules applied are as follows: \itemize{
\item if epecea positive and bfpa, stx1, and stx2 are negative then positive
\item missing if bfpa, stx1, or stx2 is missing
\item negative otherwise}
}
