% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{binarize}
\alias{binarize}
\title{Convert Yes/No to 1's and 0's}
\usage{
binarize(x)
}
\arguments{
\item{x}{a column in a dataframe}
}
\value{
the column as a binary representation.
}
\description{
Convert variables with a "Yes" or "No" entry into 1's and 0's.
Makes things easy for co-occurance matrices and math down the road.
}
