% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{binarize_summations}
\alias{binarize_summations}
\title{Binarize using dplyr functions.}
\usage{
binarize_summations(x)
}
\arguments{
\item{x}{the column passed by summarise_at as a vector from the column in the df}
}
\value{
the boolean (as 1's and 0's) vector representation is x > 0
}
\description{
Base function called with \code{summarise_at}, to convert the summations to 1 or zero. This is used for
intergrating pathogen incidence for a subject over the whole study. Without this binarisation, simply summing the incidence
can lead to sampling error.
}
\seealso{
dplyr::summarise_at
}
