% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{build_network}
\alias{build_network}
\title{For building a plotly network from the CI tables.  Needs to be written exactly as in the the edge_rewiring.Rmd programs
Apadted from the plotly for R tutorials.}
\usage{
build_network(edge_df, my_title = "")
}
\description{
For building a plotly network from the CI tables.  Needs to be written exactly as in the the edge_rewiring.Rmd programs
Apadted from the plotly for R tutorials.
}
