% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maled_tools.R
\name{clean_maled}
\alias{clean_maled}
\title{Clean the MAL-ED dataset}
\usage{
clean_maled(raw_df)
}
\arguments{
\item{raw_df:}{the dataframe directly from the}
}
\value{
: The mal-ed dataset. The
}
\description{
Parent function for importing the mal-ed. Imports that dataframe ant
cleans the dataset for only the relvent data.
}
