% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{collapse_array}
\alias{collapse_array}
\title{Collapse a 3 dimensional array to a 2 dimensional matrix.}
\usage{
collapse_array(my_array, method = "same_timepoint_binary")
}
\arguments{
\item{my_array}{The 3 dimensional array, with rows for kids kids columns equal
to the number of pathogens and slices (z) equal to the number of timepoints,}

\item{method}{the type of collapse the following are available.
     same_timepoint_binary: Do the two pathoges accor togther in the same timepoint at any time?
     TODO: Should we include same age?
     TODO: Maybe one in front, one in the back?
returns: the collapsed 2D matrix. with the same dinemsions as the DFs of lists.}
}
\description{
To collapse a 3 dimensional array into a single matrix depending on
the preferred method. Uses to build the confidence intervals and convert the
df with lists in them to rewiriable values.
}
