% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{compare_degree_dist}
\alias{compare_degree_dist}
\title{Compare the degree distributions between two matrices. Uses the get_deg_distr
function to find the distr. Then returns the distributions which are not the
same.}
\usage{
compare_degree_dist(matrix1, matrix2, by = "row")
}
\description{
matrix1: matirx of nXm
matrix2: matrix of the same dimensions
}
\details{
returns: a vector of the degree that are mismatched.
}
