% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{condense}
\alias{condense}
\title{To condense the dataframe to a list.
x: the grouped variable
return: the dataframe with a list of all the occurances}
\usage{
condense(x)
}
\description{
To condense the dataframe to a list.
x: the grouped variable
return: the dataframe with a list of all the occurances
}
