% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{condense_to_lists}
\alias{condense_to_lists}
\title{Converts the qPCR cycle numbers to an exponential value. Accoring to the
Platts mills paper, every change in cyle corresponds with a doubling in
in the relative amount of amplicon. To do this with the qCPR data, we'll
set it to 2 to the paower of the difference minus 1}
\usage{
condense_to_lists(df, qpcr_cols, fun)
}
\description{
df: the dataframe to convert
returns: The dataframe with all the qpcr columns converted to the continuous column.
     values are set to the gemotric mean.
}
