% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maled_tools.R
\name{convert_maled_binary}
\alias{convert_maled_binary}
\title{Convert MAL-ED to binary matrix}
\usage{
convert_maled_binary(df, ignore_cols, method = "simple")
}
\arguments{
\item{df:}{the dataframe of the MAL-ED dataset}

\item{ignore_cols:}{columns to ignore for binarisation, these are ID columns such as ID and
age that you do not want to mask.}
}
\value{
the MAL-ED dataframe as a binary representation.

TODO: This should be in the mal-ed only toolbox, since it is only
used for the mal-ed dataset
}
\description{
Converts the Mal-ed qpcr Data depending if they are below the threhsold value.
Reight now defaults calls binarize_qpcr() using a dplyr transmute_all function.
}
