% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{eaec_definition}
\alias{eaec_definition}
\title{The specific rule set to apply to EAEC}
\usage{
eaec_definition(aaic, aata, threshold)
}
\value{
a vector of numeric values in the range of [0,35] with
missing values to serve as the column for eaec
}
\description{
Based on the results of AAIC and AATA on the TAQ Card, implements
the rules described in 'Ecoli defintions.docx' file.
}
\details{
The rules applied:
\enumerate{
\item If AAIC and AATA is NA: return NA
\item Set EAEC as the lowest or non-missing qPCR value.
\item Negative (Ct = 35) if both AAIC and AATA are both missing.}
}
