% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{etec_definition}
\alias{etec_definition}
\title{Apply ETEC ruleset to MAL-ED}
\usage{
etec_definition(eteclt, STh, STp, threshold)
}
\arguments{
\item{eteclt}{: column containing qPCR results for eteclt}

\item{STh}{: column containing qPCR results for STh}

\item{STp}{: column containing qPCR results for STp}
}
\value{
numeric vector in range [0,35] with missing values to
serve as the column for ETEC
}
\description{
Apply the ruleset the MAL-ED dataset using STh
STp and eteclt markers to synchroise the MAL-ED with PROVIDE
}
\details{
the rulesets applied are: \itemize{
\item positive if any are positive, use lowest Ct
\item negative if all are negative or missing,
\item missing if all are missing
}
}
