% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{filter_out_nonsymmetric}
\alias{filter_out_nonsymmetric}
\title{Filter out redundant paris from symmetric matrix.}
\usage{
filter_out_nonsymmetric(d_f)
}
\arguments{
\item{d_f:}{The dataframe of the results from the configuration model, containing the actual,
the mean, and the condifence interval.}
}
\value{
A dataframe with the redundant pairs removed.
}
\description{
Iterate through the upper triangle in a symmetric matrix and remove those
entries from the results tables. Removes redudant pairs. Performs a row-wise deletion.
}
