% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{filter_out_redundancies}
\alias{filter_out_redundancies}
\title{Filter out some of thepathogen pair redundancies}
\usage{
filter_out_redundancies(d_f)
}
\description{
@description Filter out the pairs that might be considered redudant, such
 as campy pan and campy c.jejuni, especially for visualisations, since many redudant
 pairs are considered significant.
}
\details{
@param d_f: Results dataframe with the results from the configuration model such as
 actual value, ensemble mean, and confidence interval.

 @return A dataframe with the rows containing the redundant pairs removed.
}
