% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_stats.R
\name{find_co_strength}
\alias{find_co_strength}
\title{Defference between observed joint probabililty and independent}
\usage{
find_co_strength(jps, marginals)
}
\description{
Finds the difference between the observed probability and the independent probability
of two events occuring at the same time according to their marginal probabilites.

TODO: CONTINUE HERE 23Mar20
}
