% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{generalize_pathogen_names}
\alias{generalize_pathogen_names}
\title{Normalise variable names for dataset.}
\usage{
generalize_pathogen_names(df, name_set_specific, name_set_general)
}
\arguments{
\item{df}{The dataframe to be renamed}

\item{name_set_specific}{A list of the names in the original dataset, passed as a column
from a dataframe.}

\item{name_set_general}{a list of the general terms to be used, passed as a column from a
dataframe}
}
\value{
The dataframe with the names set to the more general terms.
}
\description{
For normalizing the names between the two datasets. Matches the
original name in the dataset with the general term provided in the CSV "key"
file.
}
\note{
IMPORTANT: The names need to be matched up in the dataframe for this to work.
If the name in the original file or the desired name of of the general variable
name changes, then that name needs to be changed in the respective source file containing the
variable name keys. See \code{\link{get_target_pathogens}} for source file name and how keys are obtained.
}
\seealso{
\code{\link{get_target_pathogens}}
}
