% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_datasets.R
\name{get_both_datasets}
\alias{get_both_datasets}
\title{Import MAL-ED and PROVIDE qPCR Data}
\source{
combining_datasets.R

NOTE: This method raiases a
}
\usage{
get_both_datasets(
  method = "simple",
  parent_dir = "../data/",
  bangladesh_only = F,
  specific_stool_type = NA
)
}
\arguments{
\item{method:}{a string indicating the type of generalisation for the dataset. The available
methods are:
\itemize{
\item simple : whether that subject had that pathogen at anytime for
either asymptomatic or symptomatic diarhhea.
\item simple_noncondensed : the original dataset, with a binary representation
for all the qPCR data, and NOT integrated over the entire dataset. This means
each row is a stoll sample, and reach column is a pathogen
\item simple_sum: number of times that subject had that pathogen  with
either symptomatic or asymptomatic diarhhea.
\item binary_list : for each subject and pathogen, a boolean vector
representing one's for the subject having that pathogen at that
timepoint.
\item qpcr_list : a list of all the qpcr scores.
}}

\item{parent_dir:}{a string as the parent directory of the file that's calling the method.
When running from a notebook, the parent directory is "../" and running from
a script, the parent directory is "" (empty)}
}
\value{
: both dataframes with only the target pathogens for both.
     Also contains that participant ID numbers, stool type, age, and collection date.
}
\description{
This is the main function for importing the main dataset. Contains three cidfferent
}
