% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maled_tools.R
\name{get_country_ids}
\alias{get_country_ids}
\title{Get the Country Variable for the maled Dataset}
\usage{
get_country_ids(parent_dir)
}
\description{
The Country variable for the mal-ed dataset is not aligned with the qPCR data
and is missing in some cases. This extracts
}
