% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{get_deg_distr}
\alias{get_deg_distr}
\title{Get the degree distribution for two matrices. USes DPLYR and returns a matrix
of the distribution.}
\usage{
get_deg_distr(matrix1, by)
}
\description{
returns: matrix of the distrbutios, with the first col the degree of the nodes
     and the second col is the counts.
}
