% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{get_general_terms}
\alias{get_general_terms}
\title{Import general variable names to use on both datasets}
\usage{
get_general_terms(parent_dir = "../data/", alt_filename)
}
\arguments{
\item{parent_dir}{String for the parent directory: defaults to \code{"../data"} which allows the function to be
used in notebooks and scripts. This means the file is in a directory called \code{data}, which is a seperate
directory from where the r notebook is kept. If you are storing in a spperate parent directory, then
enter that in place of the default. If the file is in the same directory, enter \code{""}. If the file is in a
parent directory, enter \code{".."}}
}
\value{
a dataframe where each row contains the variable's general terms, the version in the maled, asymptomatic,
and symptomatic PROVIDE datasets.
}
\description{
Imports the file: \code{provide_pathogen_names.csv}, which contains the
symptomatic and asymptomatic pathogen names. To be used on the respective
datasets.
}
