% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_stats.R
\name{get_joint_probability}
\alias{get_joint_probability}
\title{Calculate the joint probability for two pathogens in the dataframe.}
\usage{
get_joint_probability(d_f, names_key)
}
\arguments{
\item{d_f:}{The dataframe as returned in the original get _both_datasets()
function}

\item{names_key:}{dataframe as the "names_key" containing all the names
of the pathogens in question.}
}
\value{
A dataframe with pathogen 1, pathogen 2, and the observed joint
probability of occuring together
}
\description{
Take the tidy representation of the dataset, meaning one
observation per pathogen per study. find the probability of two
pathogens showing up together out of all kids for their respective group.
}
