% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{get_ks_df}
\alias{get_ks_df}
\title{Non-normality testing. using the kolmogornov smirnov test to analyze the normality of a distribution
of copathogen interactions.}
\usage{
get_ks_df(configuration_file, parent_dir = "../")
}
\description{
Non-normality testing. using the kolmogornov smirnov test to analyze the normality of a distribution
of copathogen interactions.
}
