% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{get_nonpathogen_cols}
\alias{get_nonpathogen_cols}
\title{Retreive the non-pathogen columns to be used as the ID columns. Can either
return the index of the ID cols or the actual labels.}
\usage{
get_nonpathogen_cols(all_cols, pathogen_cols, type = "index")
}
\description{
all-cols: all the names of the dataframe
pathogen_cols: the names of the cols containing qPCR data
type: "index" or "label" if "index" returns the index of the columns. If
     "label", returns the column names
returns: a vector of the indices or labels depending on the type specified.
}
