% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{get_pathogen_names}
\alias{get_pathogen_names}
\title{To get the pagen names for the dataframe, m is dataset containing only the pathogen
names.}
\usage{
get_pathogen_names(m)
}
\description{
To get the pagen names for the dataframe, m is dataset containing only the pathogen
names.
}
