% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{get_target_pathogens}
\alias{get_target_pathogens}
\title{Get pathogen names.}
\source{
toolkit.R
}
\usage{
get_target_pathogens(parent_dir = "../data/")
}
\value{
dataframe containinng the names for asymptomatic provide data in one column, symptomatic in another,
     and the mal-ed
}
\description{
Get the target pathogen names as indicated by the Pathogen_List.xlsx file
located in the copathogen analysis folder. Gets a dataframe with the pathogen names needed for both anaylsis.
}
\details{
The dataframe contains five columns: One column contains variable names for the PROVIDE Diarrhea dataset, the second
provides variable names for the asymptomatic dataset, and subsequent columns provide variable names for the MAL-ED dataset,
general terms to be used across all studies, and finally column of 1's and 0's what indicate whether
this variable is a pathogen or not. Each row is for a specific variable. This dataframe is used to match specific names in the
respective dataset with the genreral term to be used across studies.
}
