% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{harmonize_pathogen_data}
\alias{harmonize_pathogen_data}
\title{Hamonize the E.coli definitions for MAL-ED and PROVIDE datasets}
\usage{
harmonize_pathogen_data(d_f, df_labels)
}
\arguments{
\item{d_f}{: The MAL-ED dataset pulled from the \code{maled_full_reduced.csv} file.}

\item{df_labels}{: a dictionary of variable names for the e.coli variables as returned
by \code{\link{maled_name_dictionary}}}
}
\value{
The dataframe with the same rules applied to e.coli qPCR variables between MAL-ED and PROVIDE
}
\description{
Function for harmonizing the pathogen data with the PROVIDE study. Takes in the MAL-ED data
frame after cleaning and adds the adjustements to match.
}
\details{
Applies the following functions to the MAL-ED dataset to hamronize with PROVIDE
\itemize{
\item \code{\link{eaec_definition}}
\item \code{\link{tepec_definition}}
\item \code{\link{atepec_definition}}
\item \code{\link{stec_definition}}
\item \code{\link{lt_etec_definition}}
\item \code{\link{st_etec_definition}}
\item \code{\link{etec_definition}}
}
}
