% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_tools.R
\name{import_provide_dataset}
\alias{import_provide_dataset}
\title{Import the qPCR data from PROVIDE}
\usage{
import_provide_dataset(parent_dir = "../data/")
}
\arguments{
\item{parent_dir}{String for the parent directory: defaults to \code{"../data"} which allows the function to be
used in notebooks and scripts. This means the file is in a directory called \code{data}, which is a seperate
directory from where the r notebook is kept. If you are storing in a spperate parent directory, then
enter that in place of the default. If the file is in the same directory, enter \code{""}. If the file is in a
parent directory, enter \code{".."}}
}
\value{
the qPCR data from PROVIDE, with both Symptomatic (Diarrhea) and Asymptomatic datasets combined into one dataset
with a variable, "study" to distibguish them
}
\description{
Import the datasets for both symptomatic and asymptomatic datasets, then append
them together with a variable sesignating them as symptomatic or asymptomatic.
}
