% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{lt_etec_definition}
\alias{lt_etec_definition}
\title{Rule set for LT-ETEC}
\usage{
lt_etec_definition(eteclt, STh, STp, threshold)
}
\arguments{
\item{eteclt}{: column containing the qPCR results for eteclt}

\item{STh}{: column containing qpCR results for STh}

\item{STp}{: column containing qPCR results for STp}
}
\value{
A nummeric vector contianing the results for LT ETEC based on these
applied rules.
}
\description{
Implement rule set used in PROVIDE and appply to the
MAL-ED dataset according to the rules provided in 'Ecoli definitions.docx'
}
\details{
the rules applied are: \itemize{
\item positive if eteclt positive and STh, STp both negative (use Cq of eteclt)
\item Negative if Eteclt negative
\item Negative if Eteclt positive and and with STh or STp positive
\item Negative if Eteclt missing and STh or STp positive
\item missing if Eteclt, STh, and STp missing}
}
