% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{make_qpcrsimple_tidy}
\alias{make_qpcrsimple_tidy}
\title{Convertes the Dataframe of the simpleversion of the qPCR Data to a 'tidy' format. Allows for easier graphing.}
\usage{
make_qpcrsimple_tidy(df_long, pathogen_names)
}
\description{
Convertes the Dataframe of the simpleversion of the qPCR Data to a 'tidy' format. Allows for easier graphing.
}
