% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{maled_name_dictionary}
\alias{maled_name_dictionary}
\title{Convert the e.coli variables to their general terms.}
\usage{
maled_name_dictionary()
}
\value{
: named list (dictionary) with the abbreviations used in the code
     and the actual names as the values.
}
\description{
Generate a dictionary with the general terms as the keys and
the sepcific variable names in MAL-ED as the values.
}
