% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_tools.R
\name{only_kids_with_both_stools}
\alias{only_kids_with_both_stools}
\title{These are all the tools need for data cleaning for the PROVIDE dataset.}
\usage{
only_kids_with_both_stools(d_f, original)
}
\description{
We only want a subset of the provide dataset with participants
}
\details{
Only get the kids from the PROVIDE dataset with both symptomatic and asymptomatic stools.
}
