% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{p_of_diarrhea}
\alias{p_of_diarrhea}
\title{Write a function for getting the probability of diarrhea for a copathogen interaction in a network.
Get a table of the significant interactions and all the pathogen interactions.
The probability of diarrhea for all the copathogen interaction pairs in the dataframe}
\usage{
p_of_diarrhea(d_f)
}
\description{
Write a function for getting the probability of diarrhea for a copathogen interaction in a network.
Get a table of the significant interactions and all the pathogen interactions.
The probability of diarrhea for all the copathogen interaction pairs in the dataframe
}
