% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{process_ci_df}
\alias{process_ci_df}
\title{Calculate the Confidence interval and Significance on Configuration Results}
\usage{
process_ci_df(ci_plot, confidence_interval = 0.95)
}
\arguments{
\item{ci_plot}{: A dataframe of containing the results from the configuration model produced
in the python code. Must contain a column containing the "average", standard deviation (stdev).}

\item{confidence_interval}{:  (optional) A floating point number (either 0.95 or 0.99) indicating
the width of the interval.}
}
\value{
The original dataframe with 5 additional columns: One indicating the higher and lower intervals, a binary
indicator of significance, the distance outside that interval, and a normalised distance in the range of [0,1].
}
\description{
Calculate the confidence interval from standard deviation and the mean of the
ensemble generated by the configuration model. Calculates the significance strength, measured
on a scale from [0,1]. Includes a binary flag of whter the actual value falls ourstide the
confidence interval. Can specify the interval to cover 99% and 95% of the distribution.
}
\details{
The confidence intervals are also calucalted within the python code for the configuration model,
but overwriting this column here allows for a little more flexibility. While the distribution is technically a poisson,
I use a normal approximation for distributions with an average > 10, and a possion Chi Sqraured tool to
find the High Density Region for a poisson distribution. Can change the level of significance to widen or narrrow
that high density interval.
}
