% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_tools.R
\name{provide_binary}
\alias{provide_binary}
\title{Convert the the dataframe to a binary representation, where a one means
the value was below the threshold while a zero means it was above the threshold.}
\usage{
provide_binary(df, path_names, limit = 35, summarise_by = "sum")
}
\description{
df: the dataframe of qPCR data
path_names: the pathogen names in the dataframe. Used to exclude ID information
    from binarization.
}
