% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{qpcr_continuous}
\alias{qpcr_continuous}
\title{SImpler function to be wrapped in the 'transmute_all' function.
Works by setting the value to 2 to the power of the cycle, minus
one, since a power of zero is equal to 1.}
\usage{
qpcr_continuous(x)
}
\description{
x: the column used by dplyr::transmte_all
returns: the converted value for the column.
}
