% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maled_tools.R
\name{reduce_mad_ed}
\alias{reduce_mad_ed}
\title{Remove uneeded columns in MAL-ED}
\usage{
reduce_mad_ed(df)
}
\arguments{
\item{df:}{the MAL-ED dataframe with all columns.}
}
\value{
MAL-ED dataframe with non-id and non pathogen columns removed.
}
\description{
Reduce the mal_ed data frame to the Ct values, the ID columns,
and remove any rows that are not designated as 'Diarrhea' Or asymptomatic.
Uses Left Join to only obervations,  that have a mtaching obervation id.
}
