% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{remove_empty_slices}
\alias{remove_empty_slices}
\title{In order to make the indices more dynamic, the matrices are
initially made to be really large. This function then removes
the uneeded columns. First by finding the first slice with all
the missing values, and removing all those indices.}
\usage{
remove_empty_slices(my_array)
}
\description{
my_array: the three dimensional array with the extra slices.
returns: the array with any slices that have all missing
     values removed.
}
