% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{st_etec_definition}
\alias{st_etec_definition}
\title{Apply ST-ETEC ruleset}
\usage{
st_etec_definition(eteclt, STh, threshold)
}
\arguments{
\item{eteclt}{: column containing qPCR results for eteclt}

\item{STh}{: column containing qPCR results for STh}
}
\value{
Numeric column in range [0,35] with missing values to be used
as the ST-ETEC column
}
\description{
Apply the ruleset for ST-ETEC  to the MAL-ED
dataset as defined in in 'Ecoli definitions.docx' file.
}
\details{
Applies the following rules: \itemize{
\item Positive if STh positive, use lower Ct
\item Positive if STh positive, eteclt negative of missing, use STh Cq
\item Negatvie if STh negative (etec positive, negative or missing)
\item Missing if STh missing}
}
