% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{stec_definition}
\alias{stec_definition}
\title{Ruleset for STEC}
\usage{
stec_definition(eae, bfpa, stx1, stx2, threshold)
}
\arguments{
\item{eae}{: column containing qPCR results measuring EPECEA}

\item{stx1}{: column containing qPCR results measuring STX1}

\item{stx2}{: column containing qPCR results measuring STX2}

\item{threshold}{: a vector with length equal to the number
of rows in another column containing the value for a negative
result (35)}

\item{bpfa}{: column containing qPCR results measuring bfpA}
}
\value{
a numeric vector in range [0,35] to be used for stec_definition
}
\description{
Apply the ruleset in 'Ecoli definitions.docx' in
the copathogen folder. Applies the ruleset to MAL-ED to so we can
merge with PROVIDE
}
\details{
applies the following rules: \itemize{
\item positive if epecea is positive and bfpA is negative and stx1 and stx2 is
negative \itemize{
\item If condition met, use the lowest Ct of stx1 and stx2, and the highest Cq
value between that number and epecea}
\item Negative if epecea negative or bfpA positive and stx1 or stx2 non-missing
\item missing if epecea or bfpA missing or stx1 and stx2 missing
}
}
