% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{symmetric_df}
\alias{symmetric_df}
\title{Take the tidy CI table and expand to a matrix, filled with the values given in target}
\usage{
symmetric_df(df, path1, path2, target)
}
\description{
Add description here.
}
