% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_definitions.R
\name{tepec_definition}
\alias{tepec_definition}
\title{tEPEC defintion}
\usage{
tepec_definition(bfpa, eae, threshold)
}
\value{
A numeric vector of values in range [0,35] with missing values
for tEPEC
}
\description{
Setting the values of TEPEC based ont he qPCR results
of bfpA and etecea
}
\details{
uses the following rules: \itemize{
\item Use bfpA is etecea is negative or missing.
\item Use minuimum of bfpA and etecea if both are present.
\item Missing if bfpA is missing
\item neagtive if bfpA is negative.}
}
