% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{vacc_ci_plot}
\alias{vacc_ci_plot}
\title{Build a CI plot with pploty for Configuration model ensemble.}
\usage{
vacc_ci_plot(
  d_f,
  path1,
  path2,
  target,
  bins,
  multiple = F,
  plot_title = NA,
  joined = F,
  n_panels = 2
)
}
\arguments{
\item{d_f:}{the dataframe generated by the CI experiment function in rewire.R}

\item{path1:}{A string that's the column name containing the names fo the first pathogen}

\item{path2:}{see path1, for the second pathogen}

\item{target:}{a string as the column header for the variable to color fill for.}

\item{multiple:}{where this is a binned timeseries, while will draw a slider bar, building a plot
for each time frame.}
}
\value{
a ggplot object, a geom tile plot with frame values to be passed to a ggplot function
}
\description{
This function takes the dataframe generated by the enfoguration model
ensemble generated in R to make a plot similar to a colored correlation matrix.
}
